requirements.txt
numpy
tensorflow
keras
Pillow
h5py